[chore] ~ move addShorebirdYamlToPubspecAssets to ShorebirdEnv
