+++ packages/shorebird_cli/lib/src/shorebird_env.dart
@@ -7,6 +7,8 @@ import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/config/shorebird_yaml.dart';
 import 'package:shorebird_cli/src/platform.dart';
 import 'package:shorebird_code_push_client/shorebird_code_push_client.dart';
+import 'package:yaml/yaml.dart';
+import 'package:yaml_edit/yaml_edit.dart';
 
 /// A reference to a [ShorebirdEnv] instance.
 final shorebirdEnvRef = create(ShorebirdEnv.new);
@@ -238,4 +240,32 @@ class ShorebirdEnv {
       // https://learn.microsoft.com/en-us/azure/devops/pipelines/build/variables?view=azure-devops&tabs=yaml
       ||
       platform.environment.containsKey('TF_BUILD');
+
+  /// Adds `shorebird.yaml` to the `assets` section of the `pubspec.yaml` file
+  static void addShorebirdYamlToPubspecAssets(File pubspecFile) {
+    final pubspecContents = pubspecFile.readAsStringSync();
+    final yaml = loadYaml(pubspecContents, sourceUrl: pubspecFile.uri) as Map;
+    final editor = YamlEditor(pubspecContents);
+    if (!yaml.containsKey('flutter') || yaml['flutter'] == null) {
+      editor.update(
+        ['flutter'],
+        {
+          'assets': ['shorebird.yaml'],
+        },
+      );
+    } else {
+      if (!(yaml['flutter'] as Map).containsKey('assets')) {
+        editor.update(['flutter', 'assets'], ['shorebird.yaml']);
+      } else {
+        final assets = (yaml['flutter'] as Map)['assets'] as List;
+        if (!assets.contains('shorebird.yaml')) {
+          editor.update(['flutter', 'assets'], [...assets, 'shorebird.yaml']);
+        }
+      }
+    }
+
+    if (editor.edits.isEmpty) return;
+
+    pubspecFile.writeAsStringSync(editor.toString());
+  }
 }
